added text box when entering app

+++ network/include/types.h
@@ -2,8 +2,9 @@
 
 #include <cstdint>
 
-enum class ReturnStatus : std::int16_t {
-  kSuccess = 0, kFailure = 1
+enum class ReturnStatus : std::int8_t {
+  kSuccess = 0,
+  kFailure = 1
 };
 
 using ClientPort = std::uint16_t;
